Feature: Permanize UUID (#10)
Ensure that an existing UUID is not overwritten if candidate's data is modified or the hash algorithm changes in the future.

Checks if the id value is present in the URL. If so, use that. If not, this is a new record.

Warning: There is an edge-case where a new random user may collide with an existing recording. Not solving for this because a real DB wouldn't have this problem.

Co-authored-by: Erik Smith <github@codeblind.io>